refactor / feat: add dom color to show attacks and misses
refactored gameLoop to move DOM modification to seperate modules
